requirements.txt
aiogram
python-dotenv
python-json-logger
